Exercise2/Activity3/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.11
six==1.16.0
